examples: move interactive chat flow to basic
