+++ examples/basic/README.md
@@ -1,132 +1,49 @@
-# A2A Basic Example
+# Basic A2A Chat Example
 
-This example demonstrates a basic implementation of the Agent-to-Agent (A2A) protocol using the trpc-a2a-go library. It consists of:
+An interactive chat example written in the `taskmanager.TaskHandle` style.
+One processor serves blocking `message/send`, streaming `message/stream`, and
+long-running task operations while the client keeps a `contextId` across turns.
 
-1. A versatile text processing agent server that supports multiple operations
-2. A feature-rich CLI client that demonstrates all the core A2A protocol APIs
+The example covers:
 
-The server is written in the `taskmanager.TaskHandle` style: the minimal-edit
-port of a v0.x TaskHandler processor, with a fully synchronous body and one
-code path serving both `message/send` and `message/stream` (see "Migrating
-from v0.x" in the repository README). For the native channel style, see
-`examples/simple`.
+- Blocking and streaming message consumption
+- `returnImmediately` kickoff for a long-running task
+- `GetTask`, `SubscribeToTask`, and `CancelTask`
+- Conversation grouping and server-side message history through `contextId`
+- Multiple artifact chunks aggregated into task snapshots
 
-## Server Features
+For the minimal raw-channel `MessageProcessor` contract, see
+[`examples/simple`](../simple). For an `input-required` continuation flow, see
+[`examples/inputrequired`](../inputrequired).
 
-The server is a text processing agent capable of:
+## Run it
 
-- Processing text in various modes: reverse, uppercase, lowercase, word count
-- Supporting both streaming and non-streaming responses
-- Demonstrating multi-turn conversations with the `input-required` state
-- Handling task cancellation
-- Creating and streaming artifacts
-
-### Running the Server
+Start the server from this directory:
 
 ```bash
-cd server
-go run main.go [options]
+go run ./server
+
+# Or choose another address.
+go run ./server -host 0.0.0.0 -port 9000
 ```
 
-Server options:
-- `--host`: Host address (default: localhost)
-- `--port`: Port number (default: 8080)
-- `--desc`: Custom agent description
-- `--no-cors`: Disable CORS headers
-- `--no-stream`: Disable streaming capability
-
-## Client Features
-
-The client is a CLI application that connects to an A2A agent and provides:
-
-- Support for both streaming and non-streaming modes
-- Interactive CLI with command history
-- Session management for contextual conversations
-- Task management (create, cancel, get)
-- Agent capability discovery
-
-### Running the Client
+In another shell, run the interactive client:
 
 ```bash
-cd client
-go run main.go [options]
+go run ./client -host localhost:8080
+
+# Consume ordinary messages through message/stream.
+go run ./client -host localhost:8080 -stream
 ```
 
-Client options:
-- `--agent`: Agent URL (default: http://localhost:8080/)
-- `--timeout`: Request timeout (default: 60s)
-- `--no-stream`: Disable streaming mode
-- `--context`: Use specific context ID (generate new if empty)
-- `--history`: Number of history messages to request (default: 0)
+Useful REPL commands: `/help`, `/long-task`, `/async-long-task`, `/gettask`,
+`/subscribe`, `/cancel`, `/new`, and `/quit`.
 
-### Client Commands
+`/long-task` starts a task with `returnImmediately` and immediately subscribes
+to its events. `/async-long-task` only starts the task, leaving `/gettask`,
+`/subscribe`, and `/cancel` to be run separately.
 
-Once the client is running, you can use the following commands:
-
-- `help`: Show help message
-- `exit`: Exit the program
-- `context [id]` / `new`: Set or generate a new context ID
-- `mode [stream|sync]`: Set interaction mode (streaming or standard)
-- `cancel [task-id]`: Cancel a task
-- `get [task-id] [history]`: Get task details
-- `card`: Fetch and display the agent's capabilities card
-
-For normal interaction, type your message and press Enter. After the agent
-suspends in `input-required`, the next message continues the same `taskId`.
-
-### Text Processing Commands
-
-The server understands the following text processing commands:
-
-- `reverse <text>`: Reverses the input text
-- `uppercase <text>`: Converts text to uppercase
-- `lowercase <text>`: Converts text to lowercase
-- `count <text>`: Counts words and characters in text
-- `multi`: Start a multi-step interaction (client continues with the same `taskId`)
-- `example`: Demonstrates input-required state
-- `help`: Shows the help message
-
-## Usage Example
-
-1. Start the server:
-   ```bash
-   cd server
-   go run main.go
-   ```
-
-2. In another terminal, start the client:
-   ```bash
-   cd client
-   go run main.go
-   ```
-
-3. Try some commands:
-   ```
-   > help
-   > reverse hello world
-   > uppercase the quick brown fox
-   > multi
-   > card
-   > mode sync
-   > lowercase TESTING LOWERCASE
-   ```
-
-## A2A Protocol Implementation
-
-This example demonstrates the following A2A protocol features:
-
-- Agent discovery via Agent Cards (/.well-known/agent-card.json)
-- Task creation using message/send and message/stream
-- Task state retrieval using tasks/get
-- Task cancellation using tasks/cancel
-- Streaming updates for long-running tasks
-- Multi-turn conversations using the `input-required` state
-- Artifact generation and streaming
-
-Multi-turn progress is stored on the suspended status message's `Metadata`
-(`step` / `mode`). On continuation the framework rolls that message into
-`ExecContext.History`, so the processor needs no process-local session map.
-
-For push notifications, see [`examples/notify`](../notify) or [`examples/jwks`](../jwks)
-(`memory.WithPushNotifications` + `push.Sender`). This basic example keeps push off
-so it stays focused on the TaskHandle processor style.
+> `http.Client.Timeout` bounds the whole response body read. For
+> `message/stream` and `SubscribeToTask`, that includes the entire SSE
+> lifetime. The example uses 60 seconds so its one-word-per-second task can
+> finish with some headroom.
