+++ examples/basic/client/main.go
@@ -4,775 +4,477 @@
 //
 // trpc-a2a-go is licensed under the Apache License Version 2.0.
 
-// Package main implements a CLI host for the A2A agent.
+// Package main is an interactive chat client for the basic A2A example server.
+// It keeps a contextID across turns so the server can attach conversation history,
+// and supports async task send / get / subscribe / cancel commands.
 package main
 
 import (
 	"bufio"
 	"context"
-	"encoding/json"
 	"flag"
 	"fmt"
-	"io"
 	"log"
-	"net/http"
 	"os"
 	"strings"
 	"time"
 
 	"trpc.group/trpc-go/trpc-a2a-go/v2/client"
 	"trpc.group/trpc-go/trpc-a2a-go/v2/protocol"
-	"trpc.group/trpc-go/trpc-a2a-go/v2/server"
 )
 
-// Config holds the application configuration.
-type Config struct {
-	AgentURL         string
-	Timeout          time.Duration
-	ForceNoStreaming bool
-	ContextID        string
-	HistoryLength    int
-}
+// longTaskMarker must match examples/basic/server.
+// Sent by /long-task to simulate a long-running server turn (async/subscribe/cancel).
+const longTaskMarker = "__long_task__"
 
-// Command types for CLI.
-const (
-	cmdExit    = "exit"
-	cmdHelp    = "help"
-	cmdContext = "context"
-	cmdMode    = "mode"
-	cmdCancel  = "cancel"
-	cmdGet     = "get"
-	cmdCard    = "card"
-	cmdNew     = "new"
-)
-
-// sendResult is the task identity/state observed from a SendMessage or stream
-// response — enough for input-required continuation without a follow-up GetTasks.
-type sendResult struct {
-	taskID string
-	state  protocol.TaskState
+// session holds REPL state across turns.
+type session struct {
+	client     *client.A2AClient
+	stream     bool
+	contextID  *string
+	lastTaskID string
 }
 
 func main() {
-	// Parse command-line flags.
-	config := parseFlags()
+	host := flag.String("host", "localhost:8080", "server address")
+	stream := flag.Bool("stream", false, "use message/stream instead of message/send")
+	flag.Parse()
 
-	// Create A2A client.
-	a2aClient, err := createClient(config)
+	a2aClient, err := client.NewA2AClient(
+		fmt.Sprintf("http://%s/", *host),
+		// NOTE: http.Client.Timeout caps the WHOLE response body read, which
+		// for message/stream / SubscribeToTask is the entire SSE lifetime.
+		// 60s covers the 30s /long-task demo with some headroom.
+		client.WithTimeout(60*time.Second),
+	)
 	if err != nil {
 		log.Fatalf("Failed to create A2A client: %v", err)
 	}
 
-	// Fetch and display agent capabilities
-	agentCard, err := fetchAgentCard(config.AgentURL)
-	if err != nil {
-		log.Printf("WARNING: Failed to fetch agent card: %v", err)
-	} else {
-		displayAgentCapabilities(agentCard)
+	mode := "message/send"
+	if *stream {
+		mode = "message/stream"
 	}
+	fmt.Printf("Connected to http://%s/ (%s)\n", *host, mode)
+	printHelp()
 
-	// Display welcome message.
-	displayWelcomeMessage(config)
-
-	// Start interactive session.
-	runInteractiveSession(a2aClient, config)
-
-	fmt.Println("Exiting CLI host.")
+	s := &session{client: a2aClient, stream: *stream}
+	if err := s.runREPL(bufio.NewScanner(os.Stdin)); err != nil {
+		log.Fatalf("Failed to read stdin: %v", err)
+	}
 }
 
-// parseFlags parses command-line flags and returns a Config.
-func parseFlags() Config {
-	var config Config
-	flag.StringVar(&config.AgentURL, "agent", "http://localhost:8080/", "Target A2A agent URL")
-	flag.DurationVar(&config.Timeout, "timeout", 60*time.Second, "Request timeout (e.g., 30s, 1m)")
-	flag.BoolVar(&config.ForceNoStreaming, "no-stream", false, "Disable streaming mode")
-	flag.StringVar(&config.ContextID, "context", "", "Use specific context ID (empty = generate new)")
-	flag.IntVar(&config.HistoryLength, "history", 0, "Number of history messages to request (0 = none)")
-	flag.Parse()
-
-	// Generate a context ID if not provided
-	if config.ContextID == "" {
-		config.ContextID = protocol.GenerateContextID()
-	}
-
-	return config
-}
-
-// createClient creates a new A2A client with the given configuration.
-func createClient(config Config) (*client.A2AClient, error) {
-	return client.NewA2AClient(config.AgentURL, client.WithTimeout(config.Timeout))
-}
-
-// fetchAgentCard retrieves the agent card from the .well-known endpoint.
-func fetchAgentCard(baseURL string) (*server.AgentCard, error) {
-	// Ensure base URL ends with "/"
-	if !strings.HasSuffix(baseURL, "/") {
-		baseURL += "/"
-	}
-
-	// Construct agent card URL
-	cardURL := baseURL + ".well-known/agent-card.json"
-
-	// Make the request
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	req, err := http.NewRequestWithContext(ctx, http.MethodGet, cardURL, nil)
-	if err != nil {
-		return nil, fmt.Errorf("failed to create request: %w", err)
-	}
-
-	resp, err := http.DefaultClient.Do(req)
-	if err != nil {
-		return nil, fmt.Errorf("failed to fetch agent card: %w", err)
-	}
-	defer resp.Body.Close()
-
-	if resp.StatusCode != http.StatusOK {
-		return nil, fmt.Errorf("unexpected status code: %d", resp.StatusCode)
-	}
-
-	// Decode the response
-	var card server.AgentCard
-	if err := json.NewDecoder(resp.Body).Decode(&card); err != nil {
-		return nil, fmt.Errorf("failed to decode agent card: %w", err)
-	}
-
-	return &card, nil
-}
-
-// displayAgentCapabilities displays the capabilities from the agent card.
-func displayAgentCapabilities(card *server.AgentCard) {
-	fmt.Println("Agent Capabilities:")
-	fmt.Printf("  Name: %s\n", card.Name)
-	if card.Description != "" {
-		fmt.Printf("  Description: %s\n", card.Description)
-	}
-	fmt.Printf("  Version: %s\n", card.Version)
-
-	// Print provider if available
-	if card.Provider != nil {
-		fmt.Printf("  Provider: %s\n", card.Provider.Organization)
-	}
-
-	// Print capabilities - handle new *bool types
-	streaming := false
-	if card.Capabilities.Streaming != nil {
-		streaming = *card.Capabilities.Streaming
-	}
-	pushNotifications := false
-	if card.Capabilities.PushNotifications != nil {
-		pushNotifications = *card.Capabilities.PushNotifications
-	}
-	stateHistory := false
-	if card.Capabilities.StateTransitionHistory != nil {
-		stateHistory = *card.Capabilities.StateTransitionHistory
-	}
-
-	fmt.Printf("  Streaming: %t\n", streaming)
-	fmt.Printf("  Push Notifications: %t\n", pushNotifications)
-	fmt.Printf("  State Transition History: %t\n", stateHistory)
-
-	// Print input/output modes
-	fmt.Printf("  Input Modes: %s\n", strings.Join(card.DefaultInputModes, ", "))
-	fmt.Printf("  Output Modes: %s\n", strings.Join(card.DefaultOutputModes, ", "))
-
-	// Print skills if available
-	if len(card.Skills) > 0 {
-		fmt.Println("  Skills:")
-		for _, skill := range card.Skills {
-			fmt.Printf("    - %s: ", skill.Name)
-			if skill.Description != nil {
-				fmt.Printf("%s\n", *skill.Description)
-			} else {
-				fmt.Println("(no description)")
-			}
-
-			if len(skill.Examples) > 0 {
-				fmt.Printf("      Examples: %s\n", strings.Join(skill.Examples, ", "))
-			}
-		}
-	}
-
-	fmt.Println(strings.Repeat("-", 60))
-}
-
-// displayWelcomeMessage prints the welcome message with connection details.
-func displayWelcomeMessage(config Config) {
-	log.Printf("Connecting to agent: %s (Timeout: %v)", config.AgentURL, config.Timeout)
-	fmt.Printf("Context ID: %s\n", config.ContextID)
-	fmt.Printf("Streaming mode: %v\n", !config.ForceNoStreaming)
-	fmt.Println("Enter text to send to the agent. Type 'help' for commands or 'exit' to quit.")
-	fmt.Println(strings.Repeat("-", 60))
-}
-
-// runInteractiveSession runs the main interactive session loop.
-func runInteractiveSession(a2aClient *client.A2AClient, config Config) {
-	reader := bufio.NewReader(os.Stdin)
-	contextID := config.ContextID
-	var lastTaskID string
-	var lastTaskState protocol.TaskState
-	var useStreaming = !config.ForceNoStreaming
-
-	// Check if input is from a pipe/redirect or interactive terminal
-	stat, err := os.Stdin.Stat()
-	isInteractive := err == nil && (stat.Mode()&os.ModeCharDevice) != 0
-
-	if !isInteractive {
-		// Non-interactive mode: process all piped input at once
-		log.Println("Running in non-interactive mode (piped input)")
-		scanner := bufio.NewScanner(os.Stdin)
-		inputs := []string{}
-
-		// Read all inputs first
-		for scanner.Scan() {
-			line := strings.TrimSpace(scanner.Text())
-			if line != "" {
-				inputs = append(inputs, line)
-			}
-		}
-
-		// Process each input
-		for i, input := range inputs {
-			log.Printf("Processing input %d/%d: %s", i+1, len(inputs), input)
-
-			// Process built-in commands
-			if handled, resetContinuation := processCommand(
-				a2aClient,
-				input,
-				&config,
-				&contextID,
-				&useStreaming,
-				lastTaskID,
-			); handled {
-				if resetContinuation {
-					lastTaskState = ""
-				}
-				continue
-			}
-
-			result := processUserInput(a2aClient, input, contextID, config, useStreaming, continueTaskID(lastTaskID, lastTaskState))
-			if result.taskID != "" {
-				lastTaskID, lastTaskState = result.taskID, result.state
-			} else {
-				lastTaskState = ""
-			}
-		}
-		return
-	}
-
-	// Interactive mode: continuous loop
-	log.Println("Running in interactive mode")
+func (s *session) runREPL(scanner *bufio.Scanner) error {
 	for {
-		// Display prompt
-		fmt.Print("> ")
-
-		input, readErr := reader.ReadString('\n')
-
-		if readErr != nil {
-			if readErr == io.EOF {
-				fmt.Println("\nExiting.")
-				break
-			}
-			log.Printf("ERROR: Failed to read input: %v", readErr)
-			continue
+		fmt.Print("You> ")
+		if !scanner.Scan() {
+			fmt.Println()
+			return scanner.Err()
 		}
-
-		input = strings.TrimSpace(input)
+		input := strings.TrimSpace(scanner.Text())
 		if input == "" {
 			continue
 		}
-
-		// Process built-in commands
-		if handled, resetContinuation := processCommand(
-			a2aClient,
-			input,
-			&config,
-			&contextID,
-			&useStreaming,
-			lastTaskID,
-		); handled {
-			if resetContinuation {
-				lastTaskState = ""
-			}
-			continue
-		}
-
-		result := processUserInput(a2aClient, input, contextID, config, useStreaming, continueTaskID(lastTaskID, lastTaskState))
-		if result.taskID != "" {
-			lastTaskID, lastTaskState = result.taskID, result.state
-		} else {
-			lastTaskState = ""
-		}
-		if lastTaskState == protocol.TaskStateInputRequired {
-			fmt.Println(strings.Repeat("-", 60))
-			fmt.Println("[Additional input required to complete this task. Continue typing.]")
+		if s.handleLine(input) {
+			return nil
 		}
 	}
 }
 
-// continueTaskID returns the task ID to attach on a follow-up message when the
-// previous round suspended in input-required; otherwise the next turn is fresh.
-func continueTaskID(lastTaskID string, lastTaskState protocol.TaskState) string {
-	if lastTaskState == protocol.TaskStateInputRequired && lastTaskID != "" {
-		return lastTaskID
-	}
-	return ""
-}
-
-// processCommand handles built-in client commands. resetContinuation reports
-// whether the command intentionally leaves the current input-required task.
-func processCommand(
-	a2aClient *client.A2AClient,
-	input string,
-	config *Config,
-	contextID *string,
-	useStreaming *bool,
-	lastTaskID string,
-) (handled, resetContinuation bool) {
-	parts := strings.Fields(input)
-	cmd := strings.ToLower(parts[0])
-
-	switch cmd {
-	case cmdExit:
-		fmt.Println("Exiting.")
-		os.Exit(0)
-		return true, true
-
-	case cmdHelp:
-		displayHelpMessage()
-		return true, false
-
-	case cmdContext:
-		if len(parts) > 1 {
-			*contextID = parts[1]
-			fmt.Printf("Context ID set to: %s\n", *contextID)
-		} else {
-			*contextID = protocol.GenerateContextID()
-			fmt.Printf("Generated new context ID: %s\n", *contextID)
-		}
-		return true, true
-
-	case cmdMode:
-		if len(parts) > 1 {
-			modeStr := strings.ToLower(parts[1])
-			if modeStr == "stream" || modeStr == "streaming" {
-				*useStreaming = true
-				fmt.Println("Switched to streaming mode.")
-			} else if modeStr == "sync" || modeStr == "standard" {
-				*useStreaming = false
-				fmt.Println("Switched to standard (non-streaming) mode.")
-			} else {
-				fmt.Printf("Unknown mode: %s. Use 'stream' or 'sync'.\n", modeStr)
-			}
-		} else {
-			fmt.Printf("Current mode: %s\n", getModeName(*useStreaming))
-			fmt.Println("Usage: mode [stream|sync]")
-		}
-		return true, false
-
-	case cmdCancel:
-		taskID := lastTaskID
-		if len(parts) > 1 {
-			taskID = parts[1]
-		}
-		if taskID == "" {
-			fmt.Println("No task ID provided or available from last request.")
-			return true, false
-		}
-		canceled := cancelTask(a2aClient, taskID, config.Timeout)
-		return true, canceled && taskID == lastTaskID
-
-	case cmdGet:
-		taskID := lastTaskID
-		if len(parts) > 1 {
-			taskID = parts[1]
-		}
-		if taskID == "" {
-			fmt.Println("No task ID provided or available from last request.")
-			return true, false
-		}
-		historyLength := config.HistoryLength
-		if len(parts) > 2 {
-			if _, err := fmt.Sscanf(parts[2], "%d", &historyLength); err != nil {
-				fmt.Printf("Invalid history length: %s. Using default: %d\n", parts[2], config.HistoryLength)
-				historyLength = config.HistoryLength
-			}
-		}
-		getTask(a2aClient, taskID, historyLength, config.Timeout)
-		return true, false
-
-	case cmdCard:
-		agentCard, err := fetchAgentCard(config.AgentURL)
-		if err != nil {
-			fmt.Printf("Failed to fetch agent card: %v\n", err)
-			return true, false
-		}
-		displayAgentCapabilities(agentCard)
-		return true, false
-
-	case cmdNew:
-		*contextID = protocol.GenerateContextID()
-		fmt.Printf("Starting a new context: %s\n", *contextID)
-		return true, true
-	}
-
-	return false, false
-}
-
-// getModeName returns a user-friendly name for the current mode.
-func getModeName(streaming bool) string {
-	if streaming {
-		return "streaming (real-time updates)"
-	}
-	return "standard (non-streaming)"
-}
-
-// displayHelpMessage shows available commands and their usage.
-func displayHelpMessage() {
-	fmt.Println("Available commands:")
-	fmt.Println("  help                     - Show this help message")
-	fmt.Println("  exit                     - Exit the program")
-	fmt.Println("  context [id]             - Set or generate a new context ID")
-	fmt.Println("  mode [stream|sync]       - Set interaction mode (streaming or standard)")
-	fmt.Println("  cancel [task-id]         - Cancel a task (uses last task ID if not specified)")
-	fmt.Println("  get [task-id] [history]  - Get task details (uses last task ID if not specified)")
-	fmt.Println("  card                     - Fetch and display the agent's capabilities card")
-	fmt.Println("  new                      - Start a new context")
-	fmt.Println("")
-	fmt.Println("For normal interaction, just type your message and press Enter.")
-	fmt.Println("After input-required, the next message continues the same taskId.")
-	fmt.Println("For push notifications, see examples/notify or examples/jwks.")
-	fmt.Println(strings.Repeat("-", 60))
-}
-
-// processUserInput handles a single user input, sends it to the agent, and processes the response.
-// When continueTaskID is non-empty, the message continues a suspended (input-required) task.
-func processUserInput(
-	a2aClient *client.A2AClient,
-	input,
-	contextID string,
-	config Config,
-	useStreaming bool,
-	continueTaskID string,
-) sendResult {
-	var taskIDPtr *string
-	if continueTaskID != "" {
-		taskIDPtr = &continueTaskID
-	}
-	message := protocol.NewMessageWithContext(
-		protocol.MessageRoleUser,
-		[]*protocol.Part{protocol.NewTextPart(input)},
-		taskIDPtr,
-		&contextID,
-	)
-
-	params := createMessageParams(message, config.HistoryLength)
-
-	if useStreaming && !config.ForceNoStreaming {
-		return handleStreamingInteraction(a2aClient, params, config)
-	}
-	return handleStandardInteraction(a2aClient, params, config)
-}
-
-// createMessageParams creates the parameters for sending a message.
-func createMessageParams(message protocol.Message, historyLength int) protocol.SendMessageParams {
-	params := protocol.SendMessageParams{
-		Message: message,
-	}
-
-	// Add configuration if needed
-	if historyLength > 0 {
-		params.Configuration = &protocol.SendMessageConfiguration{
-			HistoryLength: &historyLength,
-		}
-	}
-
-	return params
-}
-
-// handleStreamingInteraction sends a streaming request to the agent and processes the response.
-func handleStreamingInteraction(
-	a2aClient *client.A2AClient,
-	params protocol.SendMessageParams,
-	config Config,
-) sendResult {
-	ctx, cancel := context.WithTimeout(context.Background(), config.Timeout*2)
-	defer cancel()
-
-	log.Printf("Sending stream request for message %s (Context: %s)...", params.Message.MessageID, *params.Message.ContextID)
-	eventChan, streamErr := a2aClient.StreamMessage(ctx, params)
-	if streamErr != nil {
-		log.Printf("ERROR: StreamMessage request failed: %v", streamErr)
-		fmt.Println(strings.Repeat("-", 60))
-		return sendResult{}
-	}
-
-	result := processStreamResponse(ctx, eventChan)
-	log.Printf("Stream processing finished for message %s", params.Message.MessageID)
-	fmt.Println(strings.Repeat("-", 60))
-	return result
-}
-
-// handleStandardInteraction sends a standard (non-streaming) request to the agent.
-func handleStandardInteraction(
-	a2aClient *client.A2AClient,
-	params protocol.SendMessageParams,
-	config Config,
-) sendResult {
-	ctx, cancel := context.WithTimeout(context.Background(), config.Timeout)
-	defer cancel()
-
-	log.Printf("Sending standard request for message %s (Context: %s)...", params.Message.MessageID, *params.Message.ContextID)
-	result, err := a2aClient.SendMessage(ctx, params)
-	if err != nil {
-		log.Printf("ERROR: SendMessage request failed: %v", err)
-		fmt.Println(strings.Repeat("-", 60))
-		return sendResult{}
-	}
-
-	fmt.Println("\n<< Agent Response:")
-	fmt.Println(strings.Repeat("-", 10))
-
-	out := sendResult{}
-	if response := result.GetMessage(); response != nil {
-		fmt.Println("  Message Response:")
-		printMessage(*response)
-	} else if response := result.GetTask(); response != nil {
-		out.taskID = response.ID
-		out.state = response.Status.State
-		fmt.Printf("  Task %s State: %s (%s)\n", response.ID, response.Status.State, formatTimestamp(response.Status.Timestamp))
-
-		if response.Status.Message != nil {
-			fmt.Println("  Message:")
-			printMessage(*response.Status.Message)
-		}
-		if len(response.Artifacts) > 0 {
-			fmt.Println("  Artifacts:")
-			for i, artifact := range response.Artifacts {
-				name := fmt.Sprintf("Artifact #%d", i+1)
-				if artifact.Name != nil {
-					name = *artifact.Name
-				}
-				fmt.Printf("    [%s]\n", name)
-				printParts(artifact.Parts)
-			}
-		}
-		if len(response.History) > 0 {
-			fmt.Println("  History:")
-			for i, msg := range response.History {
-				role := "User"
-				if msg.Role == protocol.MessageRoleAgent {
-					role = "Agent"
-				}
-				fmt.Printf("    [%d] %s:\n", i+1, role)
-				printParts(msg.Parts)
-			}
-		}
-		if response.Status.State == protocol.TaskStateInputRequired {
-			fmt.Println("  [Additional input required]")
-		}
-	} else {
-		fmt.Println("  Unknown response type")
-	}
-
-	fmt.Println(strings.Repeat("-", 60))
-	return out
-}
-
-// processStreamResponse processes the stream of events from the agent.
-func processStreamResponse(
-	ctx context.Context, eventChan <-chan protocol.StreamResponse,
-) sendResult {
-	fmt.Println("\n<< Agent Response Stream:")
-	fmt.Println(strings.Repeat("-", 10))
-
-	out := sendResult{}
-	for {
-		select {
-		case <-ctx.Done():
-			log.Printf("ERROR: Context timeout or cancellation while waiting for stream events: %v", ctx.Err())
-			return out
-
-		case event, ok := <-eventChan:
-			if !ok {
-				log.Println("Stream channel closed.")
-				if ctx.Err() != nil {
-					log.Printf("Context error after stream close: %v", ctx.Err())
-				}
-				return out
-			}
-
-			if e := event.GetMessage(); e != nil {
-				fmt.Println("  [Message Response:]")
-				printMessage(*e)
-			} else if e := event.GetTask(); e != nil {
-				out.taskID = e.ID
-				out.state = e.Status.State
-				fmt.Printf("  [Task %s State: %s (%s)]\n", e.ID, e.Status.State, formatTimestamp(e.Status.Timestamp))
-				if e.Status.Message != nil {
-					printMessage(*e.Status.Message)
-				}
-			} else if e := event.GetStatusUpdate(); e != nil {
-				out.taskID = e.TaskID
-				out.state = e.Status.State
-				fmt.Printf("  [Status Update: %s (%s)]\n", e.Status.State, formatTimestamp(e.Status.Timestamp))
-				if e.Status.Message != nil {
-					printMessage(*e.Status.Message)
-				}
-				if e.Status.State == protocol.TaskStateInputRequired {
-					fmt.Println("  [Additional input required]")
-					return out
-				}
-				if e.Final {
-					log.Printf("Final status received: %s", e.Status.State)
-					switch e.Status.State {
-					case protocol.TaskStateCompleted:
-						fmt.Println("  [Task completed successfully]")
-					case protocol.TaskStateFailed:
-						fmt.Println("  [Task failed]")
-					case protocol.TaskStateCanceled:
-						fmt.Println("  [Task was canceled]")
-					}
-					return out
-				}
-			} else if e := event.GetArtifactUpdate(); e != nil {
-				out.taskID = e.TaskID
-				name := getArtifactName(e.Artifact)
-				fmt.Printf("  [Artifact Update: %s]\n", name)
-				printParts(e.Artifact.Parts)
-				if e.LastChunk != nil && *e.LastChunk {
-					log.Printf("Final artifact received with ID %s", e.Artifact.ArtifactID)
-				}
-			} else {
-				log.Println("Warning: Received unknown stream event")
-			}
-		}
-	}
-}
-
-// getArtifactName returns the name of an artifact or a default if name is nil
-func getArtifactName(artifact protocol.Artifact) string {
-	if artifact.Name != nil {
-		return *artifact.Name
-	}
-	return fmt.Sprintf("Artifact %s", artifact.ArtifactID)
-}
-
-// cancelTask attempts to cancel a running task and reports whether the request succeeded.
-func cancelTask(a2aClient *client.A2AClient, taskID string, timeout time.Duration) bool {
-	ctx, cancel := context.WithTimeout(context.Background(), timeout)
-	defer cancel()
-
-	log.Printf("Attempting to cancel task %s...", taskID)
-
-	task, err := a2aClient.CancelTasks(ctx, protocol.TaskIDParams{ID: taskID})
-
-	if err != nil {
-		log.Printf("ERROR: Failed to cancel task %s: %v", taskID, err)
-		fmt.Printf("Failed to cancel task: %v\n", err)
+// handleLine processes one REPL line. Returns true when the session should exit.
+func (s *session) handleLine(input string) (quit bool) {
+	if strings.EqualFold(input, "help") || strings.EqualFold(input, "/help") {
+		printHelp()
 		return false
 	}
-
-	fmt.Println("Task cancellation result:")
-	fmt.Printf("  State: %s (%s)\n", task.Status.State, formatTimestamp(task.Status.Timestamp))
-
-	if task.Status.Message != nil {
-		fmt.Println("  Message:")
-		printMessage(*task.Status.Message)
+	if cmd, arg, ok := splitCommand(input); ok {
+		return s.handleCommand(cmd, arg)
 	}
-	return true
+	s.sendText(input)
+	return false
 }
 
-// getTask fetches and displays a task's current state.
-func getTask(a2aClient *client.A2AClient, taskID string, historyLength int, timeout time.Duration) {
-	ctx, cancel := context.WithTimeout(context.Background(), timeout)
-	defer cancel()
-
-	log.Printf("Fetching task %s...", taskID)
-
-	params := protocol.TaskQueryParams{ID: taskID}
-	if historyLength > 0 {
-		params.HistoryLength = &historyLength
+func (s *session) handleCommand(cmd, arg string) (quit bool) {
+	switch cmd {
+	case "/quit", "/exit":
+		fmt.Println("Bye.")
+		return true
+	case "/new":
+		s.contextID = nil
+		fmt.Println("(new conversation)")
+	case "/long-task", "/async-long-task":
+		s.runLongTask(cmd, arg)
+	case "/gettask":
+		s.withTaskID(arg, getTask)
+	case "/subscribe":
+		s.withTaskID(arg, subscribeTask)
+	case "/cancel":
+		s.withTaskID(arg, cancelTask)
+	default:
+		fmt.Printf("Unknown command: %s\n", cmd)
 	}
+	return false
+}
 
-	task, err := a2aClient.GetTasks(ctx, params)
-
+func (s *session) runLongTask(cmd, arg string) {
+	// Simulate a long-running task (/long-task also auto-subscribes).
+	if arg != "" {
+		fmt.Printf("Note: %s takes no argument; extra text is ignored\n", cmd)
+	}
+	taskID, nextContextID, err := chatAsync(s.client, longTaskMarker, s.contextID)
 	if err != nil {
-		log.Printf("ERROR: Failed to get task %s: %v", taskID, err)
-		fmt.Printf("Failed to get task: %v\n", err)
+		fmt.Printf("Error: %v\n", err)
 		return
 	}
-
-	fmt.Println("Task details:")
-	displayFinalTaskState(task)
-}
-
-// displayFinalTaskState displays the final state of a task.
-func displayFinalTaskState(task *protocol.Task) {
-	fmt.Printf("  State: %s (%s)\n", task.Status.State, formatTimestamp(task.Status.Timestamp))
-
-	if task.Status.Message != nil {
-		fmt.Println("  Message:")
-		printMessage(*task.Status.Message)
+	s.applyResult(taskID, nextContextID)
+	if taskID == "" {
+		return
 	}
-
-	if len(task.Artifacts) > 0 {
-		fmt.Println("  Artifacts:")
-		for i, artifact := range task.Artifacts {
-			name := fmt.Sprintf("Artifact #%d", i+1)
-			if artifact.Name != nil {
-				name = *artifact.Name
-			}
-			fmt.Printf("    [%s]\n", name)
-			printParts(artifact.Parts)
-		}
+	if cmd == "/async-long-task" {
+		fmt.Println("Tip: /subscribe to watch chunks, /gettask for snapshot, /cancel to stop")
+		return
 	}
-
-	if task.History != nil && len(task.History) > 0 {
-		fmt.Println("  History:")
-		for i, msg := range task.History {
-			role := "User"
-			if msg.Role == protocol.MessageRoleAgent {
-				role = "Agent"
-			}
-			fmt.Printf("    [%d] %s:\n", i+1, role)
-			printParts(msg.Parts)
-		}
+	if err := subscribeTask(s.client, taskID); err != nil {
+		fmt.Printf("Error: %v\n", err)
 	}
 }
 
-// printMessage prints the parts contained within a message.
-func printMessage(message protocol.Message) {
-	printParts(message.Parts)
-}
-
-// printParts iterates through and prints different message/artifact part types.
-func printParts(parts []*protocol.Part) {
-	for _, part := range parts {
-		printPart(part)
-	}
-}
-
-// printPart prints a single part with proper indentation.
-func printPart(part *protocol.Part) {
-	const indent = "    "
-	if text := part.TextContent(); text != "" {
-		fmt.Println(indent + text)
-	} else {
-		fmt.Printf("%s[Non-text Part: %T]\n", indent, part.Content)
-	}
-}
-
-// formatTimestamp attempts to parse and reformat an ISO8601 timestamp.
-func formatTimestamp(ts string) string {
-	if ts == "" {
-		return "(no timestamp)"
-	}
-	t, err := time.Parse(time.RFC3339, ts)
+func (s *session) withTaskID(arg string, fn func(*client.A2AClient, string) error) {
+	taskID, err := resolveTaskID(arg, s.lastTaskID)
 	if err != nil {
-		log.Printf("Warning: could not parse timestamp '%s': %v", ts, err)
-		return ts
+		fmt.Println(err)
+		return
 	}
-	return t.Local().Format(time.Stamp)
+	if err := fn(s.client, taskID); err != nil {
+		fmt.Printf("Error: %v\n", err)
+	}
+}
+
+func (s *session) sendText(input string) {
+	msg := protocol.NewMessageWithContext(
+		protocol.MessageRoleUser,
+		[]*protocol.Part{protocol.NewTextPart(input)},
+		nil,
+		s.contextID,
+	)
+	var (
+		taskID        string
+		nextContextID *string
+		err           error
+	)
+	if s.stream {
+		taskID, nextContextID, err = chatStream(s.client, msg)
+	} else {
+		taskID, nextContextID, err = chatSend(s.client, msg)
+	}
+	if err != nil {
+		fmt.Printf("Error: %v\n", err)
+		return
+	}
+	s.applyResult(taskID, nextContextID)
+}
+
+func (s *session) applyResult(taskID string, nextContextID *string) {
+	if taskID != "" {
+		s.lastTaskID = taskID
+	}
+	if nextContextID != nil {
+		s.contextID = nextContextID
+	}
+}
+
+func printHelp() {
+	fmt.Println("Commands:")
+	fmt.Println("  <text>              send a message (blocking or streaming mode)")
+	fmt.Println("  /long-task          long-running demo (async + auto-subscribe)")
+	fmt.Println("  /async-long-task    long-running demo (returnImmediately only)")
+	fmt.Println("  /gettask [id]       get task snapshot")
+	fmt.Println("  /subscribe [id]     subscribe task event stream")
+	fmt.Println("  /cancel [id]        cancel task")
+	fmt.Println("  /new                clear contextID")
+	fmt.Println("  /help               show this help")
+	fmt.Println("  /quit               exit")
+	fmt.Println()
+}
+
+// splitCommand parses "/cmd arg..." ; returns ok=false for plain text.
+func splitCommand(input string) (cmd, arg string, ok bool) {
+	if !strings.HasPrefix(input, "/") {
+		return "", "", false
+	}
+	parts := strings.SplitN(input, " ", 2)
+	cmd = strings.ToLower(parts[0])
+	if len(parts) == 2 {
+		arg = strings.TrimSpace(parts[1])
+	}
+	return cmd, arg, true
+}
+
+func resolveTaskID(arg, lastTaskID string) (string, error) {
+	if arg != "" {
+		return arg, nil
+	}
+	if lastTaskID != "" {
+		return lastTaskID, nil
+	}
+	return "", fmt.Errorf("no task id; usage needs an id or a prior /async-long-task")
+}
+
+// chatSend sends one turn via message/send and prints the final reply.
+func chatSend(c *client.A2AClient, msg protocol.Message) (taskID string, contextID *string, err error) {
+	resp, err := c.SendMessage(context.Background(), protocol.SendMessageParams{Message: msg})
+	if err != nil {
+		return "", nil, err
+	}
+
+	switch {
+	case resp.GetTask() != nil:
+		task := resp.GetTask()
+		printTaskSnapshot(task)
+		if task.ContextID != "" {
+			return task.ID, &task.ContextID, nil
+		}
+		return task.ID, nil, nil
+	case resp.GetMessage() != nil:
+		reply := resp.GetMessage()
+		fmt.Printf("message: %s\n", joinText(reply.Parts))
+		return "", reply.ContextID, nil
+	default:
+		fmt.Println("(empty response)")
+	}
+	return "", nil, nil
+}
+
+// chatAsync sends with returnImmediately and prints the immediate task snapshot.
+func chatAsync(c *client.A2AClient, text string, contextID *string) (taskID string, nextContextID *string, err error) {
+	returnImmediately := true
+	msg := protocol.NewMessageWithContext(
+		protocol.MessageRoleUser,
+		[]*protocol.Part{protocol.NewTextPart(text)},
+		nil,
+		contextID,
+	)
+	resp, err := c.SendMessage(context.Background(), protocol.SendMessageParams{
+		Message: msg,
+		Configuration: &protocol.SendMessageConfiguration{
+			ReturnImmediately: &returnImmediately,
+		},
+	})
+	if err != nil {
+		return "", nil, err
+	}
+	task := resp.GetTask()
+	if task == nil {
+		if reply := resp.GetMessage(); reply != nil {
+			fmt.Printf("message: %s\n", joinText(reply.Parts))
+			return "", reply.ContextID, nil
+		}
+		return "", nil, fmt.Errorf("expected an immediate task snapshot")
+	}
+	fmt.Printf("async task: id=%s state=%s\n", task.ID, task.Status.State)
+	printTaskSnapshot(task)
+	if task.ContextID != "" {
+		return task.ID, &task.ContextID, nil
+	}
+	return task.ID, nil, nil
+}
+
+// chatStream sends one turn via message/stream and prints events as they arrive.
+func chatStream(c *client.A2AClient, msg protocol.Message) (taskID string, contextID *string, err error) {
+	events, err := c.StreamMessage(context.Background(), protocol.SendMessageParams{Message: msg})
+	if err != nil {
+		return "", nil, err
+	}
+	return drainEvents(events)
+}
+
+func getTask(c *client.A2AClient, taskID string) error {
+	task, err := c.GetTasks(context.Background(), protocol.TaskQueryParams{ID: taskID})
+	if err != nil {
+		return err
+	}
+	fmt.Printf("task snapshot: id=%s state=%s artifacts=%d\n",
+		task.ID, task.Status.State, len(task.Artifacts))
+	printTaskSnapshot(task)
+	return nil
+}
+
+func subscribeTask(c *client.A2AClient, taskID string) error {
+	// Successful SubscribeToTask already emits the current Task snapshot as the
+	// first SSE event, then live deltas. Terminal tasks are rejected by the
+	// server — fall back to GetTasks so the user still sees the final state.
+	events, err := c.ResubscribeTask(context.Background(), protocol.TaskIDParams{ID: taskID})
+	if err != nil {
+		fmt.Printf("Subscribe unavailable: %v\n", err)
+		fmt.Println("Falling back to GetTask...")
+		return getTask(c, taskID)
+	}
+	fmt.Printf("subscribed %s\n", taskID)
+	_, _, err = drainEvents(events)
+	if err != nil {
+		return err
+	}
+	return nil
+}
+
+func cancelTask(c *client.A2AClient, taskID string) error {
+	task, err := c.CancelTasks(context.Background(), protocol.TaskIDParams{ID: taskID})
+	if err != nil {
+		return err
+	}
+	fmt.Printf("cancel result: id=%s state=%s\n", task.ID, task.Status.State)
+	printTaskSnapshot(task)
+	return nil
+}
+
+// streamPrinter formats SubscribeToTask / message/stream events for the REPL.
+type streamPrinter struct {
+	taskID        string
+	contextID     *string
+	liveOpen      bool
+	artifactNames map[string]string
+}
+
+func drainEvents(events <-chan protocol.StreamResponse) (taskID string, contextID *string, err error) {
+	p := &streamPrinter{artifactNames: map[string]string{}}
+	for event := range events {
+		p.handle(event)
+	}
+	p.endLive()
+	return p.taskID, p.contextID, nil
+}
+
+func (p *streamPrinter) handle(event protocol.StreamResponse) {
+	switch {
+	case event.GetStatusUpdate() != nil:
+		p.onStatus(event.GetStatusUpdate())
+	case event.GetArtifactUpdate() != nil:
+		p.onArtifact(event.GetArtifactUpdate())
+	case event.GetMessage() != nil:
+		p.onMessage(event.GetMessage())
+	case event.GetTask() != nil:
+		p.onTask(event.GetTask())
+	}
+}
+
+func (p *streamPrinter) endLive() {
+	if p.liveOpen {
+		fmt.Println()
+		p.liveOpen = false
+	}
+}
+
+func (p *streamPrinter) noteIDs(taskID, contextID string) {
+	if taskID != "" {
+		p.taskID = taskID
+	}
+	if contextID != "" {
+		cid := contextID
+		p.contextID = &cid
+	}
+}
+
+func (p *streamPrinter) onStatus(su *protocol.TaskStatusUpdateEvent) {
+	p.endLive()
+	p.noteIDs(su.TaskID, su.ContextID)
+	fmt.Printf("  status: %s\n", su.Status.State)
+	if su.Status.Message != nil && su.Final {
+		fmt.Printf("  status message: %s\n", joinText(su.Status.Message.Parts))
+	}
+}
+
+func (p *streamPrinter) onArtifact(au *protocol.TaskArtifactUpdateEvent) {
+	p.noteIDs(au.TaskID, au.ContextID)
+	name := rememberArtifactName(p.artifactNames, au.Artifact)
+	text := joinText(au.Artifact.Parts)
+	if text == "" {
+		return
+	}
+	if !p.liveOpen {
+		fmt.Printf("  artifact[%s]> ", name)
+		p.liveOpen = true
+	}
+	fmt.Print(text)
+	_ = os.Stdout.Sync() // show each word promptly (stdout is line-buffered)
+	if au.LastChunk != nil && *au.LastChunk {
+		fmt.Println()
+		p.liveOpen = false
+	}
+}
+
+func (p *streamPrinter) onMessage(reply *protocol.Message) {
+	p.endLive()
+	fmt.Printf("  message: %s\n", joinText(reply.Parts))
+	if reply.ContextID != nil {
+		p.contextID = reply.ContextID
+	}
+}
+
+func (p *streamPrinter) onTask(task *protocol.Task) {
+	p.endLive()
+	p.taskID = task.ID
+	p.noteIDs(task.ID, task.ContextID)
+	// First SubscribeToTask frame: metadata + seed the live line with
+	// whatever text is already aggregated, then continue with deltas.
+	fmt.Printf("  snapshot: id=%s state=%s\n", task.ID, task.Status.State)
+	if task.Status.Message != nil {
+		if text := joinText(task.Status.Message.Parts); text != "" {
+			fmt.Printf("  status message: %s\n", text)
+		}
+	}
+	for _, a := range task.Artifacts {
+		name := rememberArtifactName(p.artifactNames, a)
+		text := joinText(a.Parts)
+		if text == "" {
+			continue
+		}
+		fmt.Printf("  artifact[%s]> %s", name, text)
+		_ = os.Stdout.Sync()
+		p.liveOpen = true
+	}
+}
+
+func rememberArtifactName(names map[string]string, a protocol.Artifact) string {
+	if a.Name != nil && *a.Name != "" {
+		names[a.ArtifactID] = *a.Name
+	}
+	if name := names[a.ArtifactID]; name != "" {
+		return name
+	}
+	if a.ArtifactID != "" {
+		return a.ArtifactID
+	}
+	return "artifact"
+}
+
+func printTaskSnapshot(task *protocol.Task) {
+	// Status.message and Artifacts are different fields — label them explicitly
+	// so a short completion note is not mistaken for the final artifact text.
+	if task.Status.Message != nil {
+		if text := joinText(task.Status.Message.Parts); text != "" {
+			fmt.Printf("  status message: %s\n", text)
+		}
+	}
+	for _, artifact := range task.Artifacts {
+		text := joinText(artifact.Parts)
+		if text == "" {
+			continue
+		}
+		fmt.Printf("  final artifact[%s]: %s\n", deref(artifact.Name), text)
+	}
+}
+
+func joinText(parts []*protocol.Part) string {
+	var b strings.Builder
+	for _, part := range parts {
+		b.WriteString(part.TextContent())
+	}
+	return b.String()
+}
+
+func deref(s *string) string {
+	if s == nil {
+		return "artifact"
+	}
+	return *s
 }
