+++ examples/simple/README.md
@@ -1,12 +1,12 @@
 # Simple A2A Example (MessageProcessor contract)
 
-A minimal interactive example written directly on the v2 `MessageProcessor`
-contract — the native channel style (for the `TaskHandle` compatibility style,
-see [`examples/basic`](../basic)). It exists to show the one idea that changed
-in v2: **you write the agent once, and the framework serves every consumption
-mode from that single code path.**
+A minimal server written directly on the v2 `MessageProcessor` contract — the
+native channel style (for the `TaskHandle` compatibility style, see
+[`examples/basic`](../basic)). It reverses text, and exists to show the one
+idea that changed in v2: **you write the agent once, and the framework serves
+every consumption mode from that single code path.**
 
-## The one processor, several ways to consume it
+## The one processor, three ways to consume it
 
 The server implements a single method:
 
@@ -16,21 +16,21 @@ func (p *simpleMessageProcessor) ProcessMessage(
 ) (<-chan protocol.StreamEvent, error)
 ```
 
-It emits events — `Working` → artifact chunks → `Completed` — and never
+It emits events — `Working` → artifact → `Completed` — on a channel and never
 branches on "is this streaming?". You are not writing a response; you are
-writing the **task's event log**. Clients then consume that same log
+writing the **task's event log**. Three clients then consume that same log
 differently:
 
-| Client mode | How | What the caller gets |
+| Demo | Endpoint | What the caller gets |
 |------|----------|----------------------|
-| blocking send | type text in the REPL | Blocks; receives the **final task snapshot** (state + artifacts). |
-| streaming | `go run ./client -stream` | Receives **every event as it happens** via `message/stream`. |
-| long-running | `/long-task` or `/async-long-task` | `returnImmediately` kickoff; subscribe for live chunks, or poll with `/gettask` / cancel with `/cancel`. |
-| pure-message | send a message with no text part | Processor replies with a plain `Message` and **no task is created**. |
+| blocking send | `message/send` | Blocks; receives the **final task snapshot** (state + artifacts). Intermediate events were persisted and fanned out, but this caller sees only the derived result. |
+| returnImmediately + poll | `message/send` (`returnImmediately`) | Returns the **first** task snapshot at once; the caller polls `GetTasks` (or resubscribes) for the terminal state. |
+| streaming | `message/stream` | Receives **every event as it happens**; the channel closes when the round ends. |
+| pure-message | `message/send` (no text part) | The processor replies with a plain `Message` and **no task is created** (lazy creation) — the union carries a Message, not a Task. |
 
-The default path fills and closes a buffered raw event channel synchronously.
-The long-task path sends on a raw channel from a goroutine so
-`returnImmediately` / `SubscribeToTask` / `CancelTasks` can observe progress.
+The server code is identical for all four. `TaskID`/`ContextID` are left empty
+on the emitted events; the framework stamps them from the `ExecContext` and
+creates the task lazily on the first task event.
 
 ## Run it
 
@@ -40,18 +40,28 @@ Start the server:
 go run ./server            # listens on localhost:8080
 ```
 
-In another shell, run the interactive client:
+In another shell, run the client (it exercises all four demos in sequence):
 
 ```bash
 go run ./client -host localhost:8080
-
-# Streaming consumption mode
-go run ./client -host localhost:8080 -stream
 ```
 
-Useful REPL commands: `/help`, `/long-task`, `/async-long-task`, `/gettask`,
-`/subscribe`, `/cancel`, `/new`, `/quit`.
+Expected client output (abridged):
+
+```
+=== message/send (non-streaming, blocking) ===
+final task: state=TASK_STATE_COMPLETED
+  artifact: !dlrow olleH
+=== message/stream (streaming) ===
+status: TASK_STATE_WORKING
+artifact: !dlrow olleH
+status: TASK_STATE_COMPLETED
+stream closed (round ended)
+=== message/send (pure-message path: no task created) ===
+message reply (no task): input message must contain text.
+```
 
 > A note on the client's `WithTimeout`: `http.Client.Timeout` bounds the whole
-> response body read, which for `message/stream` / `SubscribeToTask` is the
-> entire SSE lifetime. The 60s here covers the long-task demo with headroom.
+> response body read, which for `message/stream` is the entire SSE lifetime. A
+> real long-running streaming agent needs a longer timeout (or none); the 30s
+> here only suits this toy example.
